split up all features from island area class
